requirements.txt
numpy==2.1.2
opencv-python==4.10.0.84
pillow==11.0.0
pillow-avif-plugin==1.4.6
